🚚 :memo: Move and document KeyDoorCraftingEnv
